Rename WebSearch to WebSearchProvider, inject keys
Collapse the WebSearch facade into WebSearchProvider so the registry sits
alongside the client the way LlmProvider sits alongside LlmClient. Providers
take their API key at construction instead of reading ENV; key resolution
lives in an interim WebSearchProvider.default that will become credential
lookup once search credentials are a managed model. The search tool resolves
the default provider and reports an unconfigured deployment as an ordinary
error result.
