+++ app/services/web_search_provider/tavily.rb
@@ -1,0 +1,24 @@
+module WebSearchProvider
+  # tavily.com — LLM-oriented search with content snippets.
+  class Tavily < Base
+    # Interim: env var the key comes from until search credentials are modeled.
+    ENV_KEY = "TAVILY_API_KEY"
+    ENDPOINT = "https://api.tavily.com/search"
+
+    private
+
+    def request(query, count)
+      http.post(
+        ENDPOINT,
+        body: { query: query, max_results: count }.to_json,
+        headers: { "Authorization" => "Bearer #{api_key}", "Content-Type" => "application/json" }
+      )
+    end
+
+    def map_results(json)
+      Array(json["results"]).map do |result|
+        Result.new(title: result["title"].to_s, url: result["url"].to_s, snippet: result["content"].to_s)
+      end
+    end
+  end
+end
