+++ test/services/llm_client/tools/web_search_test.rb
@@ -4,39 +4,56 @@ class LlmClient::Tools::WebSearchTest < ActiveSupport::TestCase
   def tool = LlmClient::Tools::WebSearch.new
 
   def result(index)
-    WebSearch::Result.new(title: "T#{index}", url: "https://#{index}.example", snippet: "s#{index}")
+    WebSearchProvider::Result.new(title: "T#{index}", url: "https://#{index}.example", snippet: "s#{index}")
   end
 
-  test "#execute should return normalized results as plain hashes" do
-    WebSearch.stub(:configured?, true) do
-      WebSearch.stub(:search, [result(1), result(2)]) do
-        payload = tool.execute(query: "ruby feeds")
+  # Stands in for a resolved WebSearchProvider: returns canned results or
+  # raises, so the tool can be exercised without a real provider.
+  class FakeProvider
+    def initialize(results: [], error: nil)
+      @results = results
+      @error = error
+    end
 
-        assert_equal [
-          { title: "T1", url: "https://1.example", snippet: "s1" },
-          { title: "T2", url: "https://2.example", snippet: "s2" }
-        ], payload[:results]
-      end
+    def search(_query, **)
+      raise @error if @error
+
+      @results
     end
   end
 
-  test "#execute should refuse a blank query without searching" do
-    WebSearch.stub(:search, ->(*) { flunk "search should not be called" }) do
+  test "#execute should return normalized results as plain hashes" do
+    provider = FakeProvider.new(results: [result(1), result(2)])
+
+    WebSearchProvider.stub(:default, provider) do
+      payload = tool.execute(query: "ruby feeds")
+
+      assert_equal [
+        { title: "T1", url: "https://1.example", snippet: "s1" },
+        { title: "T2", url: "https://2.example", snippet: "s2" }
+      ], payload[:results]
+    end
+  end
+
+  test "#execute should refuse a blank query without resolving a provider" do
+    WebSearchProvider.stub(:default, ->(*) { flunk "default should not be resolved" }) do
       assert_match(/Refused/, tool.execute(query: "  ")[:error])
     end
   end
 
   test "#execute should report when no provider is configured" do
-    WebSearch.stub(:configured?, false) do
-      assert_equal "Web search is not configured.", tool.execute(query: "ruby feeds")[:error]
+    unconfigured = -> { raise WebSearchProvider::ConfigurationError, "no web search provider configured" }
+
+    WebSearchProvider.stub(:default, unconfigured) do
+      assert_equal "no web search provider configured", tool.execute(query: "ruby feeds")[:error]
     end
   end
 
   test "#execute should surface provider errors as an error result" do
-    WebSearch.stub(:configured?, true) do
-      WebSearch.stub(:search, ->(*, **) { raise WebSearch::ProviderError, "Serper: HTTP 429" }) do
-        assert_equal "Serper: HTTP 429", tool.execute(query: "ruby feeds")[:error]
-      end
+    provider = FakeProvider.new(error: WebSearchProvider::ProviderError.new("Serper: HTTP 429"))
+
+    WebSearchProvider.stub(:default, provider) do
+      assert_equal "Serper: HTTP 429", tool.execute(query: "ruby feeds")[:error]
     end
   end
 end
